Fixed all errors, project done. - Buta

+++ DAT154 Oblig 2/DAT154 Oblig 2 Part 2/SpaceObjectLib/SpaceSim.cs
@@ -28,7 +28,7 @@ namespace SpaceSim
         {
             return String.Format("Name: {0}, Color: {1}, Radius {2} ", this.name, this.color, this.radius);
         }
-        public virtual double[] updatePosition(int time, SpaceObject spaceObject)
+        public virtual double[] updatePosition(double time, SpaceObject spaceObject)
         {
             return this.position;
         }
@@ -84,16 +84,17 @@ namespace SpaceSim
             Console.Write("Planet: ");
             base.Draw();
         }
-        public override double[] updatePosition(int time, SpaceObject spaceObject)
+        public override double[] updatePosition(double time, SpaceObject spaceObject)
         {
             double offsetX = 0;
             double offsetY = 0;
-
+            /*
             if (spaceObject != null)
             {
                 offsetX = spaceObject.position[0] + spaceObject.radius;
                 offsetY = spaceObject.position[1] + spaceObject.radius;
             }
+            */
 
             this.position[0] = offsetX + Math.Cos(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
             this.position[1] = offsetY + Math.Sin(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
@@ -137,17 +138,17 @@ namespace SpaceSim
             Console.Write("Moon : ");
             base.Draw();
         }
-        public override double[] updatePosition(int time, SpaceObject spaceObject)
+        public override double[] updatePosition(double time, SpaceObject spaceObject)
         {
             double offsetX = 0;
             double offsetY = 0;
-
+            /*
             if (spaceObject != null)
             {
                 offsetX = spaceObject.position[0] + spaceObject.radius;
                 offsetY = spaceObject.position[1] + spaceObject.radius;
             }
-
+            */
             this.position[0] = offsetX + Math.Cos(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
             this.position[1] = offsetY + Math.Sin(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
 
@@ -171,17 +172,17 @@ namespace SpaceSim
             Console.Write("Dwarf planet : ");
             base.Draw();
         }
-        public override double[] updatePosition(int time, SpaceObject spaceObject)
+        public override double[] updatePosition(double time, SpaceObject spaceObject)
         {
             double offsetX = 0;
             double offsetY = 0;
-
+            /*
             if (spaceObject != null)
             {
                 offsetX = spaceObject.position[0] + spaceObject.radius;
                 offsetY = spaceObject.position[1] + spaceObject.radius;
             }
-
+            */
             this.position[0] = offsetX + Math.Cos(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
             this.position[1] = offsetY + Math.Sin(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
 
@@ -235,17 +236,17 @@ namespace SpaceSim
             Console.Write("Asteroid : ");
             base.Draw();
         }
-        public override double[] updatePosition(int time, SpaceObject spaceObject)
+        public override double[] updatePosition(double time, SpaceObject spaceObject)
         {
             double offsetX = 0;
             double offsetY = 0;
-
+            /*
             if (spaceObject != null)
             {
                 offsetX = spaceObject.position[0] + spaceObject.radius;
                 offsetY = spaceObject.position[1] + spaceObject.radius;
             }
-
+            */
             this.position[0] = offsetX + Math.Cos(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
             this.position[1] = offsetY + Math.Sin(time * orbitalSpeed * 3.1416 / 180) * orbitalRadius;
 
@@ -260,4 +261,3 @@ namespace SpaceSim
         }
     }
 }
-
